ODAA-190: Added GeoJSON helper
